requirements.txt
# Dependencias principales
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==23.0.0

# Procesamiento de datos
pandas==2.3.1
numpy==2.3.2
openpyxl==3.1.5
pyreadstat==1.3.0

# Utilidades
python-multipart==0.0.20
python-dateutil==2.9.0.post0

# Testing
pytest==8.0.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Desarrollo y calidad de código
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Logging y monitoreo
structlog==24.1.0

# Validación de datos
pydantic==2.11.7
email-validator==2.1.0

# Utilidades adicionales
click==8.2.1
pathlib2==2.3.7
chardet==5.2.0
jinja2==3.1.6
requests==2.32.4
pyyaml==6.0.2